Function able to detect and extract floating point

+++ src/CommandCode.h
@@ -14,7 +14,7 @@ struct StoreCMD{
 typedef struct Variable Variable;
 struct Variable {
   char name;
-  int integer;
+  double integer;
   int fraction;
   int isValid;
 };
